+++ src/common/interfaces/exceptions/exception-filter-metadata.interface.ts
@@ -1,5 +1,5 @@
 import { ExceptionFilter } from './exception-filter.interface';
-import { Metatype } from './metatype.interface';
+import { Metatype } from '../metatype.interface';
 
 export interface ExceptionFilterMetadata {
     func: ExceptionFilter['catch'];