+++ src/common/interfaces/exceptions/index.ts
@@ -1,0 +1,2 @@
+export * from './exception-filter-metadata.interface';
+export * from './exception-filter.interface'