+++ src/common/interfaces/middlewares/middlewares-consumer.interface.ts
@@ -1,4 +1,4 @@
-import { Metatype } from './metatype.interface';
+import { Metatype } from '../metatype.interface';
 import { MiddlewareConfigProxy } from './middleware-config-proxy.interface';
 
 export interface MiddlewaresConsumer {