+++ src/core/index.ts
@@ -1,4 +1,13 @@
+/*
+ * Nest @core
+ * Copyright(c) 2017-... Kamil Mysliwiec
+ * www.nestjs.com || www.kamilmysliwiec.com
+ * MIT Licensed
+ */
+
 export { HttpException } from './exceptions/http-exception';
-export { NestMiddleware, MiddlewareConfiguration } from './middlewares/interfaces';
 export { MiddlewareBuilder } from './middlewares/builder';
 export { ModuleRef } from './injector/module-ref';
+export * from './nest-factory';
+export * from './nest-application';
+export * from './nest-microservice';