+++ src/core/injector/injector.ts
@@ -1,14 +1,14 @@
 import 'reflect-metadata';
 import { InstanceWrapper } from './container';
-import { UnknownDependenciesException } from '../../errors/exceptions/unknown-dependencies.exception';
-import { RuntimeException } from '../../errors/exceptions/runtime.exception';
+import { UnknownDependenciesException } from '../errors/exceptions/unknown-dependencies.exception';
+import { RuntimeException } from '../errors/exceptions/runtime.exception';
 import { Module } from './module';
-import { Metatype } from '../../common/interfaces/metatype.interface';
-import { Controller } from '../../common/interfaces/controller.interface';
-import { Injectable } from '../../common/interfaces/injectable.interface';
+import { Metatype } from '@nestjs/common/interfaces/metatype.interface';
+import { Controller } from '@nestjs/common/interfaces/controllers/controller.interface';
+import { Injectable } from '@nestjs/common/interfaces/injectable.interface';
 import { MiddlewareWrapper } from '../middlewares/container';
-import { isUndefined, isNil, isFunction } from '../../common/utils/shared.utils';
-import { PARAMTYPES_METADATA, SELF_DECLARED_DEPS_METADATA } from '../../common/constants';
+import { isUndefined, isNil, isFunction } from '@nestjs/common/utils/shared.utils';
+import { PARAMTYPES_METADATA, SELF_DECLARED_DEPS_METADATA } from '@nestjs/common/constants';
 
 export class Injector {
     public loadInstanceOfMiddleware(
@@ -20,9 +20,9 @@ export class Injector {
         const currentMetatype = collection.get(metatype.name);
         if (currentMetatype.instance !== null) return;
 
-        this.resolveConstructorParams(wrapper as any, module, null, argsInstances => {
+        this.resolveConstructorParams(wrapper as any, module, null, null, (instances) => {
             collection.set(metatype.name, {
-                instance: new metatype(...argsInstances),
+                instance: new metatype(...instances),
                 metatype,
             });
         });
@@ -45,12 +45,12 @@ export class Injector {
         });
     }
 
-    public loadInstanceOfComponent(wrapper: InstanceWrapper<Injectable>, module: Module) {
+    public loadInstanceOfComponent(wrapper: InstanceWrapper<Injectable>, module: Module, context: Module[] = []) {
         const components = module.components;
-        this.loadInstance<Injectable>(wrapper, components, module);
+        this.loadInstance<Injectable>(wrapper, components, module, context);
     }
 
-    public loadInstance<T>(wrapper: InstanceWrapper<T>, collection, module: Module) {
+    public loadInstance<T>(wrapper: InstanceWrapper<T>, collection, module: Module, context: Module[] = []) {
         const { metatype, name, inject } = wrapper;
         const currentMetatype = collection.get(name);
         if (isUndefined(currentMetatype)) {
@@ -58,14 +58,14 @@ export class Injector {
         }
 
         if (currentMetatype.isResolved) return;
-        this.resolveConstructorParams<T>(wrapper, module, inject, argsInstances => {
+        this.resolveConstructorParams<T>(wrapper, module, inject, context, (instances) => {
             if (isNil(inject)) {
                 currentMetatype.instance = Object.assign(
                     currentMetatype.instance,
-                    new metatype(...argsInstances),
+                    new metatype(...instances),
                 );
             } else {
-                currentMetatype.instance = currentMetatype.metatype(...argsInstances);
+                currentMetatype.instance = currentMetatype.metatype(...instances);
             }
             currentMetatype.isResolved = true;
         });
@@ -75,11 +75,20 @@ export class Injector {
         wrapper: InstanceWrapper<T>,
         module: Module,
         inject: any[],
+        context: Module[],
         callback: (args) => void) {
 
+        let isResolved = true;
         const args = isNil(inject) ? this.reflectConstructorParams(wrapper.metatype) : inject;
-        const instances = args.map(param => this.resolveSingleParam<T>(wrapper, param, module));
-        callback(instances);
+
+        const instances = args.map((param) => {
+            const paramWrapper = this.resolveSingleParam<T>(wrapper, param, module, context);
+            if (paramWrapper.isExported && !paramWrapper.isResolved) {
+                isResolved = false;
+            }
+            return paramWrapper.instance;
+        });
+        isResolved && callback(instances);
     }
 
     public reflectConstructorParams<T>(type: Metatype<T>): any[] {
@@ -94,7 +103,12 @@ export class Injector {
         return Reflect.getMetadata(SELF_DECLARED_DEPS_METADATA, type) || [];
     }
 
-    public resolveSingleParam<T>(wrapper: InstanceWrapper<T>, param: Metatype<any> | string | symbol, module: Module) {
+    public resolveSingleParam<T>(
+        wrapper: InstanceWrapper<T>,
+        param: Metatype<any> | string | symbol,
+        module: Module,
+        context: Module[]) {
+
         if (isUndefined(param)) {
             throw new RuntimeException();
         }
@@ -102,32 +116,66 @@ export class Injector {
             module,
             isFunction(param) ? (param as Metatype<any>).name : param,
             wrapper,
+            context,
         );
     }
 
-    public resolveComponentInstance<T>(module: Module, name: any, wrapper: InstanceWrapper<T>) {
+    public resolveComponentInstance<T>(module: Module, name: any, wrapper: InstanceWrapper<T>, context: Module[]) {
         const components = module.components;
-        const instanceWrapper = this.scanForComponent(components, name, module, wrapper);
+        const instanceWrapper = this.scanForComponent(components, name, module, wrapper, context);
 
-        if (!instanceWrapper.isResolved) {
+        if (!instanceWrapper.isResolved && !instanceWrapper.isExported) {
             this.loadInstanceOfComponent(components.get(name), module);
-        }
-        return instanceWrapper.instance;
-    }
-
-    public scanForComponent(components: Map<string, any>, name: any, module: Module, { metatype }) {
-        if (components.has(name)) {
-            return components.get(name);
-        }
-
-        const instanceWrapper = this.scanForComponentInRelatedModules(module, name);
-        if (isNil(instanceWrapper)) {
-            throw new UnknownDependenciesException(metatype.name);
         }
         return instanceWrapper;
     }
 
-    public scanForComponentInRelatedModules(module: Module, name: any) {
+    public scanForComponent(components: Map<string, any>, name: any, module: Module, { metatype }, context: Module[] = []) {
+        const component = this.scanForComponentInScopes(context, name, metatype);
+        if (component) {
+            return component;
+        }
+        return components.has(name) ? components.get(name) :
+                this.scanForComponentInExports(components, name, module, metatype, context);
+    }
+
+    public scanForComponentInExports(components: Map<string, any>, name: any, module: Module, { metatype }, context: Module[] = []) {
+        const instanceWrapper = this.scanForComponentInRelatedModules(module, name, context);
+        if (!isNil(instanceWrapper)) {
+            return instanceWrapper;
+        }
+        const { exports } = module;
+        if (!exports.has(metatype.name)) {
+            throw new UnknownDependenciesException(metatype.name);
+        }
+        return {
+            instance: null,
+            isResolved: false,
+            isExported: true,
+        };
+    }
+
+    public scanForComponentInScopes(context: Module[], name: any, metatype) {
+        context = context || [];
+        for (const ctx of context) {
+            const component = this.scanForComponentInScope(ctx, name, metatype);
+            if (component) return component;
+        }
+        return null;
+    }
+
+    public scanForComponentInScope(context: Module, name: any, metatype) {
+        try {
+            return this.scanForComponent(
+                context.components, name, context, { metatype }, null,
+            );
+        }
+        catch (e) {
+            return null;
+        }
+    }
+
+    public scanForComponentInRelatedModules(module: Module, name: any, context: Module[]) {
         const relatedModules = module.relatedModules || [];
         let component = null;
 
@@ -137,7 +185,7 @@ export class Injector {
 
             component = components.get(name);
             if (!component.isResolved) {
-                this.loadInstanceOfComponent(component, relatedModule);
+                this.loadInstanceOfComponent(component, relatedModule, [].concat(context, module));
             }
         });
         return component;