+++ src/core/middlewares/container.ts
@@ -1,6 +1,6 @@
-import { MiddlewareConfiguration } from './interfaces/middleware-configuration.interface';
-import { NestMiddleware } from './interfaces/nest-middleware.interface';
-import { Metatype } from '../../common/interfaces/metatype.interface';
+import { MiddlewareConfiguration } from '@nestjs/common/interfaces/middlewares/middleware-configuration.interface';
+import { NestMiddleware } from '@nestjs/common/interfaces/middlewares/nest-middleware.interface';
+import { Metatype } from '@nestjs/common/interfaces/metatype.interface';
 
 export class MiddlewaresContainer {
     private readonly middlewares = new Map<string, Map<string, MiddlewareWrapper>>();