+++ src/core/router/interfaces/explorer.inteface.ts
@@ -1,6 +1,6 @@
-import { Controller } from '../../../common/interfaces/index';
-import { Metatype } from '../../../common/interfaces/metatype.interface';
+import { Controller } from '@nestjs/common/interfaces/index';
+import { Metatype } from '@nestjs/common/interfaces/metatype.interface';
 
 export interface RouterExplorer {
-     explore(instance: Controller, metatype: Metatype<Controller>, moduleName: string);
+    explore(instance: Controller, metatype: Metatype<Controller>, moduleName: string);
 }