+++ src/microservices/index.ts
@@ -1,3 +1,11 @@
+/*
+ * Nest @microservices
+ * Copyright(c) 2017-... Kamil Mysliwiec
+ * www.nestjs.com || www.kamilmysliwiec.com
+ * MIT Licensed
+ */
+
 export * from './utils';
 export * from './interfaces';
 export * from './client';
+export * from './enums';